Refactor to match eslint and react standards.
