+++ .eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  "extends": "airbnb",
+  "env": { "browser": true },
+  "plugins": [
+    "react",
+    "jsx-a11y",
+    "import"
+  ]
+};
