+++ .gitignore
@@ -1,5 +1,8 @@
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
+# build/dist
+/build
+
 # dependencies
 /node_modules
