--- src/Components/Home.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Home = () => (
-  <div>
-    <h2>Home</h2>
-  </div>
-)
-
-export default Home
