--- src/Components/Routes.js
@@ -1,34 +1,0 @@
-// import React from 'react'
-//
-// import {
-//   BrowserRouter as Router,
-//   Route,
-//   Link,
-//   Switch,
-//   Redirect
-// } from 'react-router-dom'
-//
-// import Header from '../Header'
-// import Topics from '../Topics'
-// import Home from '../Home'
-// import About from '../About'
-// import NoMatch from '../NoMatch'
-//
-// const App = () => (
-//   <Router>
-//     <div>
-//
-//       <Header />
-//
-//       <Switch>
-//         <Route path="/" exact component={Home}/>
-//         <Route path="/about" component={About}/>
-//         <Route path="/topics" component={Topics}/>
-//         <Route component={NoMatch}/>
-//       </Switch>
-//
-//     </div>
-//   </Router>
-// )
-//
-// export default Routes
