--- src/Components/Topic.js
@@ -1,8 +1,0 @@
-import React from 'react'
-const Topic = ({ match }) => (
-  <div>
-    <h3>{match.params.topicId}</h3>
-  </div>
-)
-
-export default Topic
