+++ src/components/Home.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Home = () => (
+  <div>
+    <h2>Home</h2>
+  </div>
+);
+
+export default Home;
