+++ src/components/SearchContainer.jsx
@@ -1,0 +1,57 @@
+import React, { Component } from 'react';
+import _ from 'lodash';
+
+import Fuse from 'fuse.js';
+import PeopleJSON from '../stubData/10000.json';
+import SearchDisplay from './SearchDisplay';
+
+const options = {
+  shouldSort: true,
+  threshold: 0.3,
+  location: 0,
+  distance: 100,
+  maxPatternLength: 32,
+  minMatchCharLength: 3,
+  keys: [
+    'participantName',
+    // 'clientId',
+    // 'employeeId',
+    // 'accountNumber',
+    // 'planAdmin'
+  ],
+};
+
+export default class SearchContainer extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      people: PeopleJSON,
+      peopleFiltered: [],
+      term: '',
+    };
+    this.handleChange = this.handleChange.bind(this);
+  }
+
+  fuseResults(term) {
+    const fuse = new Fuse(this.state.people, options);
+    const filteredResults = fuse
+      .search(term)
+      .slice(0, 11);
+    this.setState({ peopleFiltered: filteredResults });
+  }
+
+  handleChange(e) {
+    const term = e.target.value;
+    this.setState({ term });
+    _.debounce(() => this.fuseResults(term), 400)();
+  }
+
+  render() {
+    return (
+      <SearchDisplay
+        peopleFiltered={this.state.peopleFiltered}
+        handleChangeCallback={this.handleChange}
+      />
+    );
+  }
+}
