+++ src/components/Topic.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Topic = ({ match }) => (
+  <div>
+    <h3>{match.params.topicId}</h3>
+  </div>
+);
+
+export default Topic;
