+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './Components/App.js';
+import App from './components/App';
 import './assets/css/core.css';
 
 ReactDOM.render(
