+++ src/routes.js
@@ -1,0 +1,3 @@
+export default routes = {
+
+};
